AzureFunctions/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

# Azure default imports
azure-functions

# Custom imports
alpaca-py
